Feat/v2 client side

+++ OpenEnding_2022.3.8f1/Assets/Scripts/Game/Connect/Connect_Scene.cs
@@ -11,7 +11,7 @@ using UnityEditor;
 public class Connect_Scene : Singleton<Connect_Scene>
 {
     public int n;
-    public int r;
+    public float r;
     
     public DeviceObject userOwnDeviceObject;
